django-stripe/pyproject.toml
[tool.poetry]
name = "django-stripe"
version = "0.1.0"
description = ""
authors = ["nasif raihan <nasif.raihan78@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.4"
djangorestframework = "^3.16.0"
stripe = "^12.3.0"
python-decouple = "^3.8"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.7.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
